Move index.js to lib directory
