+++ lib/index.js
@@ -38,7 +38,7 @@ function BambooHR(options) {
           password: 'x'
         },
         headers: {
-            'User-Agent': 'node-bamboohr/' + require('./package.json').version + ' (node/' + process.version + ')'
+            'User-Agent': 'node-bamboohr/' + require('../package.json').version + ' (node/' + process.version + ')'
         }
     })
 }
